+++ crates/store/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 0.2.0-git
+
+### Major
+
+- Remove `!` (never type) `Storage` implementation
+
 ## 0.1.1
 
 ### Minor